# optimized brigand::is_set