--- a/brigand/algorithms/is_set.hpp
+++ b/brigand/algorithms/is_set.hpp
@@ -6,45 +6,41 @@
 =================================================================================================**/
 #pragma once
 
-#include <brigand/sequences/make_sequence.hpp>
-#include <brigand/types/integer.hpp>
+#include <initializer_list>
+
+#include <brigand/sequences/range.hpp>
 #include <brigand/types/type.hpp>
 
 namespace brigand
 {
 namespace detail
 {
-  template<class Ints, class...T>
-  struct is_set_impl
-  {
-    template<class U>
-    static std::true_type is_set(U);
-
-    static std::true_type is_set();
-  };
-
-  template<class, class T>
-  struct unique_x_t
+  template<int, class T> struct unique_x_t
+  // :type_<T> // best with g++
   { operator type_<T> (); };
 
-  template<class... T>
-  struct unique_set : T... {};
+  template<class Ints, class... Ts>
+  struct is_set_impl;
 
-  template<class Int, class... Ints, class L, class... R>
-  struct is_set_impl<list<Int, Ints...>, L, R...>
+  template<>
+  struct is_set_impl<list<>>
   {
-    template<class U, class = decltype(static_cast<type_<U>>(unique_set<unique_x_t<Int, L>, unique_x_t<Ints, R>...>()))>
-    static std::false_type is_set(type_<U>);
+    using type = std::true_type;
+  };
 
-    template<class U>
-    static decltype(is_set_impl<list<Ints...>, R...>::is_set(type_<L>()))
-    is_set(U);
+  template<int... Ints, class... Ts>
+  struct is_set_impl<list<std::integral_constant<int, Ints>...>, Ts...>
+  {
+    struct Pack : unique_x_t<Ints, Ts>... {};
 
-    static decltype(is_set_impl<list<Ints...>, R...>::is_set(type_<L>()))
-    is_set(...);
+    template<class... Us>
+    static auto is_set(Us...) -> decltype(std::initializer_list<int>{(static_cast<Us>(Pack()), 1)...}, std::true_type{});
+    static std::false_type is_set(...);
+
+    using type = decltype(is_set(type_<Ts>()...));
   };
 }
 
-  template <class... T>
-  using is_set = decltype(detail::is_set_impl<make_sequence<uint32_t<0>, sizeof...(T)>, T...>::is_set());
+  template<class... Ts>
+  using is_set = typename detail::is_set_impl<range<int, 0, sizeof...(Ts)>, Ts...>::type;
 }