--- a/brigand/algorithms/is_set.hpp
+++ b/brigand/algorithms/is_set.hpp
@@ -6,8 +6,6 @@
 =================================================================================================**/
 #pragma once
 
-#include <initializer_list>
-
 #include <brigand/sequences/range.hpp>
 #include <brigand/types/type.hpp>