`//list<int, int, int, ...>`
`using list = brigand::wrap<brigand::filled_list<int, N_L1>, brigand::is_set>;`

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 37M | 00.11s - 40M | 00.22s - 46M | 00.56s - 66M | ko
gcc | 00.06s - 21M | 00.12s - 31M | 00.29s - 52M | 01.09s - 122M | 05.95s - 373M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 36M | 00.06s - 36M | 00.06s - 37M | 00.07s - 38M | 00.08s - 39M
gcc | 00.04s - 19M | 00.04s - 19M | 00.05s - 21M | 00.07s - 23M | 00.12s - 28M

----

`//list<_1, _2, _3, ...>`
`using list = brigand::wrap<brigand::range<int, 0, N_L1>, brigand::is_set>;`

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 37M | 00.14s - 40M | 00.40s - 47M | 02.14s - 67M | ko
gcc | 00.06s - 22M | 00.13s - 33M | 00.32s - 56M | 01.22s - 135M | 06.46s - 417M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 35M | 00.06s - 36M | 00.09s - 37M | 00.17s - 39M | 00.49s - 43M
gcc | 00.04s - 19M | 00.06s - 22M | 00.10s - 28M | 00.23s - 49M | 00.73s - 122M

----

`//list<int, int, ..., _1, _2, _3, ...>`
`using list = brigand::wrap<brigand::append<brigand::filled_list<int, N_L1/2>, brigand::range<int, 0, N_L1/2>>, brigand::is_set>;`

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 37M | 00.12s - 40M | 00.23s - 46M | 00.69s - 66M | ko
gcc | 00.06s - 22M | 00.13s - 32M | 00.31s - 54M | 01.14s - 126M | 06.17s - 386M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 35M | 00.06s - 36M | 00.06s - 37M | 00.07s - 38M | 00.09s - 40M
gcc | 00.04s - 19M | 00.05s - 20M | 00.05s - 22M | 00.08s - 25M | 00.13s - 31M